Change the api_secret field to api_id.
The application.api_secret won't really be a secret at all.
It'll just be an extra identifier to check against.

e.g. A user might have authorized an application, and now that
application might be embedded via javascript on a page using the
will-someday-exist javascript API.

The page with the javascript API will publically have the
app's api_id, and Breakpoint will have set a cookie in the user's
browser. These two things combined will authenticate the user.