Allow the API to be documented via GitHub wiki. See below for info.
This is experimental, and might be totally a bad way to do this.
The goal here is that we can make API documentation a community thing.
If users want to post neat examples of how to use our API in cool ways,
then let them.

:watermelon:

So we off-load documentation to the Frontend's wiki on github.
If the page exists on the wiki, we redirect to it.
If not, we error (and in the future, probably give a link that says
"if you know what you're doing, feel free to create this wiki page.").

Some problems with this method:
* The API is important - if someone evil updates the wiki, people will
  be confused and not know how to use the API.
* If GitHub goes down for some reason, they bring our API documentation
  down with them.
* Due to the way we check if a page exists, if GitHub changes workflow for
  not allowing logged-out visitors to edit wiki pages, it will break us.
  Well, actually, worst case, we'll just never error -- we'll redirect to
  an empty dialog (the blank wiki page textarea).